add pipelinecalc and move all checkers calcilations from channelcalc to checkers
